Move vectorised call up in structure
The call to "berechne_relative_einstrahlflaeche" was too complicated and depended on certain inputs. By decoupling its vectorised structure from its original function (i.e. to calculate relative irridaiton area) it is easier to call the function as intended. To get summarised results, a mapped funciton is created in the optimisation.